Настраивает модуляцию
